Array dup check BST built.
Duplicate array entry added and functioning normally along with sorting.

I think the tree is generating correctly, I fed the same array through some webapp visualizers and it matched some of them.

Although not sure why it rarely has a right node, even if it makes sense that it does. Testing the code of the highest rated example has same output so I am going to assume I got it right but I wouldn't be surprised if need to change it but good enough to start working on other functions.