Refactored buildTree(), find() function added.
As I thought on the prior commit I needed to rework things, it much cleaner now and confirmed in the Discord that the output is what expect it to be. It has the clean split if input 7 numbers.

Also added the find() function that returns the node that has the value that is input.