Depth(), Height(), , isBalanced(), and depth value on nodes added.
Height() finds how far to the furthest leaf of the tree, Depth() finds how deep a specific node is.

isBalanced() compares the height of the left and right branches and reports if they more than 1 height apart.

After much effort I finally gave up trying to generate depth in function and went for the much easier option of just having it stored in the node when created.